Bugfix: Don't delete pictures when editing the previous character
The problem was the translation of "new index" to "original index"

With the character immediately before an image, "new index" points to the beginning of the replacement range.

The old logic said, "When you're trying to find the upper bound of a range that overlaps a replacement range, use the end of the replacement range."

However, this is only true when trying to find the upper bound **and** you're looking at something **in the middle** of a replacement range. If you're looking at the beginning, you can keep looking at the beginning.