Add Kindle Highlights importer (#48)
This is based on running javascript in an in-app browser -- this is the first time I've used this technique. I'm using Typescript to make it easier to write safe Javascript. It's easy to imagine a world where there are lots of web scrapers implemented in Javascript delivered over the air.