Bugfix: Let books appear multiple times in the list
If you read the same book in multiple years, it needs to be able to show up multiple times. It would crash in this scenario before this bugfix because items in a diffable data source need to be unique.

To add uniqueness, I now add "year read" to the item.